+++ packages/typespec-client-generator-core/src/tsp-index.ts
@@ -16,6 +16,7 @@ import {
   $flattenProperty,
   $legacyHierarchyBuilding,
   $markAsLro,
+  $nextLinkVerb,
   $operationGroup,
   $override,
   $paramAlias,
@@ -58,5 +59,6 @@ export const $decorators = {
     hierarchyBuilding: $legacyHierarchyBuilding,
     flattenProperty: $flattenProperty,
     markAsLro: $markAsLro,
+    nextLinkVerb: $nextLinkVerb,
   } satisfies AzureClientGeneratorCoreLegacyDecorators,
 };